set up the reader ship api, the methods are similar to what is being used in the book count, some adjustments to teh setup of the files.

+++ entities/structs.go
@@ -14,11 +14,11 @@ type BookCount struct {
 
 // Readership is a struct that holds the number of readers for a given language
 type Readership struct {
-	Country    string `json:"country"`
-	Isocode    string `json:"isocode"`
-	Books      int    `json:"books"`
-	Authors    int    `json:"authors"`
-	Readership int    `json:"readership"`
+	Country    string  `json:"country"`
+	Isocode    string  `json:"isocode"`
+	Books      int     `json:"books"`
+	Authors    int     `json:"authors"`
+	Readership float64 `json:"readership"`
 }
 
 // Status is a struct that holds the status of the service
